+++ src/moveOrder.hpp
@@ -1,6 +1,7 @@
 #ifndef MOVEORDER_H
 #define MOVEORDER_H
 
+#include "common.hpp"
 #include "gameState.hpp"
 
 class MoveOrder {
@@ -10,10 +11,10 @@ class MoveOrder {
         reset();
     }
 
-
     void cutoff(Depth depth, unsigned move) {
         killerMove[depth] = move;
     }
+
     void reset() {
         for(unsigned i=0; i<DEPTH_MAX; i++) {
             killerMove[i] = MOVE_INVALID;
@@ -23,8 +24,7 @@ class MoveOrder {
     void reorder(Depth depth, unsigned cachedMove, unsigned moves[WIDTH]);
 
   private:
-    unsigned killerMove[DEPTH_MAX];
-
+    uint8 killerMove[DEPTH_MAX];
 };
